reorganized the library and readerNotes code
