+++ models/Reader.js
@@ -8,13 +8,7 @@ const { Job } = require('./Job')
 const { Tag } = require('./Tag')
 const { Attribution } = require('./Attribution')
 const { urlToId } = require('../utils/utils')
-const urlparse = require('url').parse
-const route = require('path-match')({
-  sensitive: false,
-  strict: false,
-  end: false
-})
-const match = route('/publication-:context/:path*')
+
 const short = require('short-uuid')
 const translator = short()
 
@@ -64,222 +58,6 @@ class Reader extends BaseModel {
     return readers[0]
   }
 
-  static async getLibraryCount (readerId, filter) {
-    let author, attribution, type
-    if (filter.author) author = Attribution.normalizeName(filter.author)
-    if (filter.attribution) {
-      attribution = Attribution.normalizeName(filter.attribution)
-    }
-    if (filter.type) {
-      type =
-        filter.type.charAt(0).toUpperCase() +
-        filter.type.substring(1).toLowerCase()
-    }
-    let resultQuery = Publication.query(Publication.knex())
-      .count()
-      .whereNull('Publication.deleted')
-      .andWhere('Publication.readerId', '=', readerId)
-
-    if (filter.title) {
-      resultQuery = resultQuery.where(
-        'Publication.name',
-        'ilike',
-        `%${filter.title.toLowerCase()}%`
-      )
-    }
-    if (filter.type) {
-      resultQuery = resultQuery.where('Publication.type', '=', type)
-    }
-    if (filter.language) {
-      resultQuery = resultQuery.whereJsonSupersetOf(
-        'Publication.metadata:inLanguage',
-        [filter.language]
-      )
-    }
-    if (filter.keyword) {
-      resultQuery = resultQuery.whereJsonSupersetOf(
-        'Publication.metadata:keywords',
-        [filter.keyword.toLowerCase()]
-      )
-    }
-    if (filter.author) {
-      resultQuery = resultQuery
-        .leftJoin(
-          'Attribution',
-          'Attribution.publicationId',
-          '=',
-          'Publication.id'
-        )
-        .where('Attribution.normalizedName', '=', author)
-        .andWhere('Attribution.role', '=', 'author')
-    }
-    if (filter.attribution) {
-      resultQuery = resultQuery
-        .leftJoin(
-          'Attribution',
-          'Attribution.publicationId',
-          '=',
-          'Publication.id'
-        )
-        .where('Attribution.normalizedName', 'like', `%${attribution}%`)
-      if (filter.role) {
-        resultQuery = resultQuery.andWhere('Attribution.role', '=', filter.role)
-      }
-    }
-    if (filter.collection) {
-      resultQuery = resultQuery
-        .leftJoin(
-          'publication_tag',
-          'publication_tag.publicationId',
-          '=',
-          'Publication.id'
-        )
-        .leftJoin('Tag', 'publication_tag.tagId', '=', 'Tag.id')
-        .where('Tag.name', '=', filter.collection)
-        .andWhere('Tag.type', '=', 'reader:Stack')
-    }
-    if (filter.search) {
-      const search = filter.search.toLowerCase()
-      resultQuery = resultQuery.where(nestedQuery => {
-        nestedQuery
-          .where('Publication.name', 'ilike', `%${search}%`)
-          .orWhere('Publication.abstract', 'ilike', `%${search}%`)
-          .orWhere('Publication.description', 'ilike', `%${search}%`)
-          .orWhereJsonSupersetOf('Publication.metadata:keywords', [search])
-      })
-    }
-
-    const result = await resultQuery
-    return result[0].count
-  }
-
-  static async getLibrary (
-    readerAuthId /*: string */,
-    limit /*: number */,
-    offset /*: number */,
-    filter /*: any */
-  ) {
-    offset = !offset ? 0 : offset
-    let author, attribution, type
-    if (filter.author) author = Attribution.normalizeName(filter.author)
-    if (filter.attribution) {
-      attribution = Attribution.normalizeName(filter.attribution)
-    }
-    if (filter.type) {
-      type =
-        filter.type.charAt(0).toUpperCase() +
-        filter.type.substring(1).toLowerCase()
-    }
-
-    const readers = await Reader.query(Reader.knex())
-      .where('Reader.authId', '=', readerAuthId)
-      .skipUndefined()
-      .eager('[tags, publications]')
-      .modifyEager('publications', builder => {
-        builder
-          .select(
-            'Publication.id',
-            'Publication.metadata',
-            'Publication.name',
-            'Publication.datePublished',
-            'Publication.status',
-            'Publication.type',
-            'Publication.encodingFormat',
-            'Publication.published',
-            'Publication.updated',
-            'Publication.deleted',
-            'Publication.resources'
-          )
-          .from('Publication')
-        builder.distinct('Publication.id')
-        builder.whereNull('Publication.deleted')
-        if (filter.title) {
-          const title = filter.title.toLowerCase()
-          builder.where('Publication.name', 'ilike', `%${title}%`)
-        }
-        if (filter.language) {
-          builder.whereJsonSupersetOf('Publication.metadata:inLanguage', [
-            filter.language
-          ])
-        }
-        if (filter.keyword) {
-          builder.whereJsonSupersetOf('Publication.metadata:keywords', [
-            filter.keyword.toLowerCase()
-          ])
-        }
-        if (filter.type) {
-          builder.where('Publication.type', '=', type)
-        }
-        builder.leftJoin(
-          'Attribution',
-          'Attribution.publicationId',
-          '=',
-          'Publication.id'
-        )
-        builder.leftJoin(
-          'publication_tag',
-          'publication_tag.publicationId',
-          '=',
-          'Publication.id'
-        )
-        builder.leftJoin('Tag', 'publication_tag.tagId', '=', 'Tag.id')
-        builder.whereNull('Tag.deleted')
-        if (filter.author) {
-          builder
-            .where('Attribution.normalizedName', '=', author)
-            .andWhere('Attribution.role', '=', 'author')
-        }
-        if (filter.attribution) {
-          builder.where(
-            'Attribution.normalizedName',
-            'like',
-            `%${attribution}%`
-          )
-          if (filter.role) {
-            builder.andWhere('Attribution.role', '=', filter.role)
-          }
-        }
-        builder.eager('[tags, attributions]')
-        if (filter.collection) {
-          builder
-            .where('Tag.name', '=', filter.collection)
-            .andWhere('Tag.type', '=', 'reader:Stack')
-        }
-        if (filter.search) {
-          const search = filter.search.toLowerCase()
-          builder.where(nestedBuilder => {
-            nestedBuilder
-              .where('Publication.name', 'ilike', `%${search}%`)
-              .orWhere('Publication.abstract', 'ilike', `%${search}%`)
-              .orWhere('Publication.description', 'ilike', `%${search}%`)
-              .orWhereJsonSupersetOf('Publication.metadata:keywords', [search])
-          })
-        }
-        if (filter.orderBy === 'title') {
-          if (filter.reverse) {
-            builder.orderBy('Publication.name', 'desc')
-          } else {
-            builder.orderBy('Publication.name')
-          }
-        } else if (filter.orderBy === 'datePublished') {
-          if (filter.reverse) {
-            builder.orderByRaw('"datePublished" NULLS FIRST')
-          } else {
-            builder.orderByRaw('"datePublished" DESC NULLS LAST')
-          }
-        } else {
-          if (filter.reverse) {
-            builder.orderBy('Publication.updated')
-          } else {
-            builder.orderBy('Publication.updated', 'desc')
-          }
-        }
-        builder.limit(limit)
-        builder.offset(offset)
-      })
-    return readers[0]
-  }
-
   static async checkIfExistsByAuthId (
     authId /*: string */
   ) /*: Promise<boolean> */ {
@@ -289,153 +67,6 @@ class Reader extends BaseModel {
       authId
     )
     return readers.length > 0
-  }
-
-  static async getNotesCount (readerId, filters) {
-    // note: not applied with filters.document
-
-    let resultQuery = Note.query(Note.knex())
-      .count()
-      .whereNull('Note.deleted')
-      .andWhere('Note.readerId', '=', readerId)
-      .leftJoin('NoteBody', 'NoteBody.noteId', '=', 'Note.id')
-
-    if (filters.publication) {
-      resultQuery = resultQuery.where(
-        'Note.publicationId',
-        '=',
-        urlToId(filters.publication)
-      )
-    }
-    if (filters.motivation) {
-      resultQuery = resultQuery.where(
-        'NoteBody.motivation',
-        '=',
-        filters.motivation
-      )
-    }
-
-    if (filters.search) {
-      resultQuery = resultQuery.where(
-        'NoteBody.content',
-        'ilike',
-        '%' + filters.search.toLowerCase() + '%'
-      )
-    }
-
-    if (filters.collection) {
-      resultQuery = resultQuery
-        .leftJoin('note_tag', 'note_tag.noteId', '=', 'Note.id')
-        .leftJoin('Tag', 'note_tag.tagId', '=', 'Tag.id')
-        .whereNull('Tag.deleted')
-        .where('Tag.name', '=', filters.collection)
-        .andWhere('Tag.type', '=', 'reader:Stack')
-    }
-
-    const result = await resultQuery
-
-    return result[0].count
-  }
-
-  static async getNotes (
-    readerAuthId /*: string */,
-    limit /*: number */,
-    offset /*: number */,
-    filters /*: any */
-  ) /*: Promise<Array<any>> */ {
-    offset = !offset ? 0 : offset
-    const { Document } = require('./Document')
-    const qb = Reader.query(Reader.knex()).where('authId', '=', readerAuthId)
-    let doc
-    if (filters.document) {
-      const path = urlparse(filters.document).path // '/publications/{pubid}/path/to/file'
-      const startIndex = path.split('/', 3).join('/').length // index of / before path/to/file
-      const docPath = path.substring(startIndex + 1) // 'path/to/file'
-      const publicationId = path.substring(14, startIndex) // {pubid}
-      doc = await Document.byPath(publicationId, docPath)
-      // $FlowFixMe
-      if (!doc) doc = { id: 'does not exist' } // to make sure it returns an empty array instead of failing
-
-      // no pagination for filter by document
-      offset = 0
-      limit = 100000
-    }
-
-    const readers = await qb
-      .eager('replies.[publication.[attributions], body]')
-      .modifyEager('replies', builder => {
-        builder.modifyEager('body', bodyBuilder => {
-          bodyBuilder.select('content', 'language', 'motivation')
-          bodyBuilder.whereNull('deleted')
-        })
-        builder.select('Note.*').from('Note')
-        builder.distinct('Note.id')
-        // load details of parent publication for each note
-        builder.modifyEager('publication', pubBuilder => {
-          pubBuilder.whereNull('Publication.deleted')
-          pubBuilder.select(
-            'id',
-            'name',
-            'type',
-            'published',
-            'updated',
-            'deleted'
-          )
-        })
-        builder.whereNull('Note.deleted')
-
-        // filters
-        if (filters.publication) {
-          builder.where('publicationId', '=', urlToId(filters.publication))
-        }
-        if (filters.document) {
-          builder.where('documentId', '=', urlToId(doc.id))
-        }
-
-        builder.leftJoin('NoteBody', 'NoteBody.noteId', '=', 'Note.id')
-        if (filters.motivation) {
-          builder.where('NoteBody.motivation', '=', filters.motivation)
-        }
-
-        if (filters.search) {
-          builder.where(
-            'NoteBody.content',
-            'ilike',
-            '%' + filters.search.toLowerCase() + '%'
-          )
-        }
-
-        builder.leftJoin('note_tag', 'note_tag.noteId', '=', 'Note.id')
-        builder.leftJoin('Tag', 'note_tag.tagId', '=', 'Tag.id')
-        builder.whereNull('Tag.deleted')
-        if (filters.collection) {
-          builder
-            .where('Tag.name', '=', filters.collection)
-            .andWhere('Tag.type', '=', 'reader:Stack')
-        }
-
-        // orderBy
-        if (filters.orderBy === 'created') {
-          if (filters.reverse) {
-            builder.orderBy('Note.published')
-          } else {
-            builder.orderBy('Note.published', 'desc')
-          }
-        }
-
-        if (filters.orderBy === 'updated') {
-          if (filters.reverse) {
-            builder.orderBy('Note.updated')
-          } else {
-            builder.orderBy('Note.updated', 'desc')
-          }
-        }
-
-        // paginate
-        builder.limit(limit).offset(offset)
-      })
-
-    return readers[0]
   }
 
   static async createReader (
